Db changes

+++ MoviesSystem/MoviesSystem.Models/Description.cs
@@ -15,6 +15,6 @@ namespace MoviesSystem.Models
         [Column(TypeName = "text")]
         public string Summary { get; set; }
 
-        public ushort Year { get; set; }
+        public int Year { get; set; }
     }
 }
